backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
tortoise-orm==0.25.1
asyncpg==0.30.0
redis==6.4.0
pydantic-settings==2.10.1
passlib==1.7.4
bcrypt==4.0.1
